requirements.txt
numpy==1.24.4
pandas
plotly-express
streamlit